project/requirements.txt
# Core dependencies
python-dotenv>=0.19.0
pydantic>=1.8.2

# Database
cassandra-driver>=3.25.0
pymilvus>=2.0.0

# AI/ML
nvidia-nemo>=1.0.0
torch>=1.9.0
transformers>=4.11.0

# Web scraping
beautifulsoup4>=4.9.3
requests>=2.26.0
firecrawl>=0.1.0

# Testing and validation
pytest>=6.2.5
docker>=5.0.0
safety>=1.10.3

# API and web
fastapi>=0.68.1
uvicorn>=0.15.0

# Utilities
python-jose>=3.3.0
passlib>=1.7.4 